refactor(client): change name styles to camelCase in all project
