+++ packages/client/src/app/app.module.css
@@ -3,7 +3,7 @@
   font-weight: 400;
   font-style: normal;
   font-family: 'Rubik Bubbles';
-  src: url('../assets/fonts/rubic-bubbles-cyrillic.woff2') format('woff2');
+  src: url('../assets/fonts/rubicBubblesCyrillic.woff2') format('woff2');
   unicode-range: U+0301, U+0400-045F, U+0490-0491, U+04B0-04B1, U+2116;
 }
 
@@ -12,7 +12,7 @@
   font-weight: 400;
   font-style: normal;
   font-family: 'Rubik Bubbles';
-  src: url('../assets/fonts/rubic-bubbles-latin.woff2') format('woff2');
+  src: url('../assets/fonts/rubicBubblesLatin.woff2') format('woff2');
   unicode-range: U+0000-00FF, U+0131, U+0152-0153, U+02BB-02BC, U+02C6, U+02DA,
     U+02DC, U+2000-206F, U+2074, U+20AC, U+2122, U+2191, U+2193, U+2212, U+2215,
     U+FEFF, U+FFFD;
@@ -34,7 +34,7 @@ body {
   font-family: 'Rubik Bubbles', fantasy;
 }
 
-.link_button {
+.linkButton {
   padding: 10px 50px;
   border: 5px solid black;
   border-radius: 15px;
@@ -48,10 +48,10 @@ body {
   transition: all 0.2s linear;
 }
 
-.link_button:hover {
+.linkButton:hover {
   background: hsl(29deg 6% 74%);
 }
 
-.link_button:active {
+.linkButton:active {
   background: hsl(30deg 76% 64%);
 }
