+++ packages/client/src/app/index.tsx
@@ -1,5 +1,4 @@
 import { AppRouting } from '../pages'
-import { NavigationBar } from '../components/navigation-bar'
 import { Provider as StoreProvider } from 'react-redux'
 import { store } from '../store'
 import { BrowserRouter } from 'react-router-dom'
@@ -9,7 +8,6 @@ const App = () => {
   return (
     <BrowserRouter>
       <StoreProvider store={store}>
-        <NavigationBar />
         <AppRouting />
       </StoreProvider>
     </BrowserRouter>
