+++ packages/client/src/assets/images/heroBomb.png
