--- packages/client/src/components/navigation-bar/index.ts
@@ -1,1 +1,0 @@
-export { NavigationBar } from './ui'
