--- packages/client/src/components/navigation-bar/navigation-bar.module.css
@@ -1,4 +1,0 @@
-.navLink {
-  color: 'teal';
-  font-family: "Rubik Bubbles", fantasy;
-}
