+++ packages/client/src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react'
 import ReactDOM from 'react-dom/client'
 import { App } from './app'
-import { ErrorBoundary } from './components/error-boundary/error-boundary'
+import { ErrorBoundary } from './components/errorBoundary/errorBoundary'
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <StrictMode>
